Implements CustomAnalyzer for title attribute.
This class allows to customize a Lucene analyzer by using a 'config' dict. It currently supports (optional) lowercase conversion, stemming and stopword removal. By default, this analyzer would be applied to all fields that require certain analysis (i.e. at least tokenization), which is not what we want. Therefore, a PerFieldAnalyzerWrapper is used to tell Lucene for which field the analyzer is meant.